requirements.txt
streamlit==1.31.0
pillow==10.2.0
pytesseract==0.3.10
requests==2.31.0
googletrans==4.0.0-rc1
pandas==2.2.0
transformers==4.37.2
datasets==2.16.1
torch==2.1.2
opencv-python==4.9.0.80 